--- a/Makefile
+++ b/Makefile
@@ -60,7 +60,7 @@ SWAGGER_IMAGE=quay.io/goswagger/swagger:v$(SWAGGER_VER)
 SWAGGER=docker run --rm -it -u ${shell id -u} -v "${PWD}:/go/src/" -w /go/src/ $(SWAGGER_IMAGE)
 SWAGGER_GEN_DIR=swagger-gen
 
-JAEGER_DOCKER_PROTOBUF=jaegertracing/protobuf:0.2.0
+JAEGER_DOCKER_PROTOBUF=jaegertracing/protobuf:0.3.0
 
 COLOR_PASS=$(shell printf "\033[32mPASS\033[0m")
 COLOR_FAIL=$(shell printf "\033[31mFAIL\033[0m")
@@ -172,6 +172,8 @@ lint-staticcheck:
 	time staticcheck ./... \
 		| grep -v \
 			-e model/model.pb.go \
+			-e proto-gen \
+			-e _test.pb.go \
 			-e thrift-gen/ \
 			-e swagger-gen/ \
 		>> $(LINT_LOG) || true