--- a/go.mod
+++ b/go.mod
@@ -75,3 +75,6 @@ require (
 	gopkg.in/yaml.v2 v2.4.0
 	honnef.co/go/tools v0.2.0
 )
+
+// TODO: remove once the 0.14.2 or 0.15.0 is released
+replace github.com/apache/thrift => github.com/jaegertracing/thrift v0.14.1-patch1