--- a/pkg/gogocodec/codec_test.go
+++ b/pkg/gogocodec/codec_test.go
@@ -17,7 +17,7 @@ package gogocodec
 import (
 	"testing"
 
-	"github.com/golang/protobuf/proto"
+	"github.com/golang/protobuf/proto" //lint:ignore SA1019 deprecated package
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/types/known/emptypb"