--- a/proto-gen/codec.go
+++ b/proto-gen/codec.go
@@ -10,6 +10,7 @@ import (
 )
 
 const jaegerProtoGenPkgPath = "github.com/jaegertracing/jaeger/proto-gen"
+const jaegerModelPkgPath = "github.com/jaegertracing/jaeger/model"
 
 func init() {
 	encoding.RegisterCodec(newCodec())
@@ -36,7 +37,7 @@ func (c *gogoCodec) Marshal(v interface{}) ([]byte, error) {
 	t := reflect.TypeOf(v)
 	elem := t.Elem()
 	// use gogo proto only for Jaeger types
-	if elem != nil && strings.HasPrefix(elem.PkgPath(), jaegerProtoGenPkgPath) {
+	if useGogo(elem) {
 		return gogoproto.Marshal(v.(gogoproto.Message))
 	}
 	return proto.Marshal(v.(proto.Message))
@@ -47,8 +48,12 @@ func (c *gogoCodec) Unmarshal(data []byte, v interface{}) error {
 	t := reflect.TypeOf(v)
 	elem := t.Elem()
 	// use gogo proto only for Jaeger types
-	if elem != nil && strings.HasPrefix(elem.PkgPath(), jaegerProtoGenPkgPath) {
+	if useGogo(elem) {
 		return gogoproto.Unmarshal(data, v.(gogoproto.Message))
 	}
 	return proto.Unmarshal(data, v.(proto.Message))
+}
+
+func useGogo(t reflect.Type) bool {
+	return t!= nil && (strings.HasPrefix(t.PkgPath(), jaegerProtoGenPkgPath) || strings.HasPrefix(t.PkgPath(), jaegerModelPkgPath))
 }