--- a/proto-gen/codec_test.go
+++ b/proto-gen/codec_test.go
@@ -1,0 +1,22 @@
+package proto_gen
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/jaegertracing/jaeger/model"
+)
+
+func TestCodec(t *testing.T) {
+	c := newCodec()
+	s1 := &model.Span{OperationName: "foo", TraceID: model.NewTraceID(1, 2)}
+	data, err := c.Marshal(s1)
+	require.NoError(t, err)
+
+	s2 := &model.Span{}
+	err = c.Unmarshal(data, s2)
+	require.NoError(t, err)
+	assert.Equal(t, s1, s2)
+}