Builds on top of #2857 

Resolves #2771
Resolves #2172