@pavolloffay 2021-06-07
> In any case, is this something we'd want for the long-term, or do you think we still need a better solution in the future?

correct, this is rather a short-term solution. Since the gogo is compatible only with go proto V1 and it will not be made compatible with v2 we will have to migrate. It seems that solutions like https://github.com/planetscale/vtprotobuf (based on https://github.com/gogo/protobuf/issues/691#issuecomment-854136033) will be way forward. It is a plugin for proto compiler that generates serialization methods, it does not modify generated model classes.
 This has a couple of implications for us:
* no support for nonnullable fields
* no support for custom types e.g. (TraceID)