@pavolloffay 2021-06-08
Benchmark results look good to me. See the results below. The benchmark code is in this PR in the `model` package.

This PR:
```
GOMAXPROCS=1 go test -bench=. -test.benchtime=10s -benchmem -cpuprofile profile_cpu.out ./model                                                                                             
goos: linux
goarch: amd64
pkg: github.com/jaegertracing/jaeger/model
cpu: Intel(R) Core(TM) i7-9850H CPU @ 2.60GHz
BenchmarkKeyValueIsLessAndEquals 	313754542	        38.84 ns/op	       0 B/op	       0 allocs/op
BenchmarkKeyValuesSort           	83939259	       154.1 ns/op	      24 B/op	       1 allocs/op
BenchmarkKeyValuesWrapping       	1000000000	         0.2424 ns/op	       0 B/op	       0 allocs/op
BenchmarkSpanHash                	  552618	     21807 ns/op	    2695 B/op	     112 allocs/op
BenchmarkBatchSerialization/marshal         	15824212	       749.9 ns/op	     480 B/op	       7 allocs/op
BenchmarkBatchSerialization/unmarshal       	 7336722	      1737 ns/op	    1592 B/op	      25 allocs/op
PASS
ok  	github.com/jaegertracing/jaeger/model	68.924s
```

Master with `github.com/gogo/protobuf/proto`
```
GOMAXPROCS=1 go test -bench=. -test.benchtime=10s -benchmem -cpuprofile profile_cpu.out ./model                                                                                               
go: downloading github.com/golang/protobuf v1.3.5
goos: linux
goarch: amd64
pkg: github.com/jaegertracing/jaeger/model
cpu: Intel(R) Core(TM) i7-9850H CPU @ 2.60GHz
BenchmarkKeyValueIsLessAndEquals 	306750854	        39.29 ns/op	       0 B/op	       0 allocs/op
BenchmarkKeyValuesSort           	84167293	       151.4 ns/op	      24 B/op	       1 allocs/op
BenchmarkKeyValuesWrapping       	1000000000	         0.2390 ns/op	       0 B/op	       0 allocs/op
BenchmarkSpanHash                	  544084	     21888 ns/op	    2695 B/op	     112 allocs/op
BenchmarkBatchSerialization/marshal         	11489486	      1047 ns/op	     624 B/op	      10 allocs/op
BenchmarkBatchSerialization/unmarshal       	 7310154	      1633 ns/op	    1592 B/op	      25 allocs/op
PASS
ok  	github.com/jaegertracing/jaeger/model	68.099s
```

Master with `ithub.com/golang/protobuf/proto`
```
GOMAXPROCS=1 go test -bench=. -test.benchtime=10s -benchmem -cpuprofile profile_cpu.out ./model                                                                                                 
goos: linux
goarch: amd64
pkg: github.com/jaegertracing/jaeger/model
cpu: Intel(R) Core(TM) i7-9850H CPU @ 2.60GHz
BenchmarkKeyValueIsLessAndEquals 	306549897	        42.74 ns/op	       0 B/op	       0 allocs/op
BenchmarkKeyValuesSort           	77749569	       158.2 ns/op	      24 B/op	       1 allocs/op
BenchmarkKeyValuesWrapping       	1000000000	         0.2391 ns/op	       0 B/op	       0 allocs/op
BenchmarkSpanHash                	  553023	     21951 ns/op	    2695 B/op	     112 allocs/op
BenchmarkBatchSerialization/marshal         	11297779	      1064 ns/op	     624 B/op	      10 allocs/op
BenchmarkBatchSerialization/unmarshal       	 7264723	      1635 ns/op	    1592 B/op	      25 allocs/op
PASS
ok  	github.com/jaegertracing/jaeger/model	68.930s
```