@pavolloffay 2021-06-08
> Do I read it right that golang/protobuf and gogo/protobuf are on the same ballpark? If this benchmark is correct, is it even worth using gogo at all in our case?

We do not use gogo only for perf but also for using custom types (e.g. `TraceID`) and minimize allocations by not using nullable fields - this is not reflected in serialization benchmark. Also even when using golang/proto serialization the gogo generated code is most likely used. To properly compare gogo and stdlib the benchmark would have to generate model twice. For purpose of this PR ignore the golang/protobuf result.